moves v3 tests to root level
