+++ tests/Http/TagsTest.php
@@ -3,8 +3,8 @@
 namespace Tests\Http;
 
 use Tests\TestCase;
+use Rogue\Models\Tag;
 use Rogue\Models\Post;
-use Rogue\Models\User;
 
 class TagsTest extends TestCase
 {
@@ -12,83 +12,173 @@ class TagsTest extends TestCase
      * Test that a POST request to /tags updates the post's tags and
      * creates a new event and tagged entry.
      *
-     * POST /tags
+     * POST /v3/posts/:post_id/tag
      * @return void
      */
-    public function testAddingATagToAPost()
+    public function testTaggingAPost()
     {
         // Create the models that we will be using
         $post = factory(Post::class)->create();
 
         // Apply the tag to the post
-        $response = $this->actingAsAdmin()->postJson('tags', [
-            'post_id' => $post->id,
+        $response = $this->withAccessToken($this->randomUserId(), 'admin')->postJson('api/v3/posts/' . $post->id . '/tags', [
             'tag_name' => 'Good Photo',
         ]);
 
-        $response->assertSuccessful();
+        $response->assertStatus(200);
 
         // Make sure that the post's tags are updated.
         $this->assertContains('Good Photo', $post->tagNames());
 
-        // Make sure we created a event for the tag.
-        $this->assertDatabaseHas('events', [
-            'eventable_type' => 'Rogue\Models\Post',
-        ]);
+        // @TODO: Make sure we created a event for the tag once events are refactored.
     }
 
     /**
-     * Test that a POST request to /tags soft-deletes an existing tag
+     * Test a POST request to /tags without the activity scope.
+     *
+     * POST /v3/posts/:post_id/tag
+     * @return void
+     */
+    public function testTaggingAPostWithoutActivityScope()
+    {
+        // Create the models that we will be using
+        $post = factory(Post::class)->create();
+
+        // Apply the tag to the post
+        $response = $this->postJson('api/v3/posts/' . $post->id . '/tags', [
+            'tag_name' => 'Good Photo',
+        ]);
+
+        $response->assertStatus(401);
+        $this->assertEquals('Unauthenticated.', $response->decodeResponseJson()['message']);
+    }
+
+    /**
+     * Test that a non-admin cannot tag a post.
+     *
+     * POST /v3/posts/:post_id/tag
+     * @return void
+     */
+    public function testUnauthenticatedUserCannotTagAPost()
+    {
+        // Create the models that we will be using
+        $post = factory(Post::class)->create();
+
+        // Apply the tag to the post
+        $response = $this->postJson('api/v3/posts/' . $post->id . '/tags', [
+            'tag_name' => 'Good Photo',
+        ]);
+
+        $response->assertStatus(401);
+    }
+
+    /**
+     * Test that a DELETE request to /tags deletes an existing tag
      * on a post, creates a new event, and tagged entry.
      *
-     * POST /tags
+     * DELETE /v3/posts/:post_id/tag
      * @return void
      */
     public function testDeleteTagOnAPost()
     {
-        $this->actingAsAdmin();
+        // @TODO: Gateway keeps the "Token" from this PHPUnit call for later,
+        // and so we always think requests are anonymous. That's no good!
+        // We can swap this back once that's fixed in Gateway.
+        // $post = factory(Post::class)->create()->tag('Good Photo');
 
-        // Create a post with a tag.
         $post = factory(Post::class)->create();
-        $post->tag('Good Photo');
 
-        $response = $this->postJson('tags', [
-            'post_id' => $post->id,
+        $this->withAccessToken($this->randomUserId(), 'admin')->postJson('api/v3/posts/' . $post->id . '/tags', [
+             'tag_name' => 'Good Photo',
+         ]);
+
+        $this->assertContains('Good Photo', $post->tagNames());
+
+        $response = $this->withAccessToken($this->randomUserId(), 'admin')->deleteJson('api/v3/posts/' . $post->id . '/tags', [
             'tag_name' => 'Good Photo',
         ]);
 
         // Make sure that the tag is deleted.
         $response->assertStatus(200);
-        $this->assertEmpty($post->tagNames());
+        $this->assertEmpty($post->fresh()->tagNames());
 
-        // Make sure we created an event for the tag.
-        $this->assertDatabaseHas('events', [
-            'eventable_type' => 'Rogue\Models\Post',
+        // @TODO: Make sure we created a event for the tag once events are refactored.
+    }
+
+    /**
+     * Test  a DELETE request without activity scope.
+     *
+     * DELETE /v3/posts/:post_id/tag
+     * @return void
+     */
+    public function testDeleteTagOnAPostWithoutActivityScope()
+    {
+        // @TODO: Gateway keeps the "Token" from this PHPUnit call for later,
+        // and so we always think requests are anonymous. That's no good!
+        // We can swap this back once that's fixed in Gateway.
+        // $post = factory(Post::class)->create()->tag('Good Photo');
+
+        $post = factory(Post::class)->create();
+
+        $response = $this->deleteJson('api/v3/posts/' . $post->id . '/tags', [
+            'tag_name' => 'Good Photo',
         ]);
+
+        $response->assertStatus(401);
+        $this->assertEquals('Unauthenticated.', $response->decodeResponseJson()['message']);
+    }
+
+    /**
+     * Test that a non-admin cannot untag a post.
+     *
+     * DELETE /v3/posts/:post_id/tag
+     * @return void
+     */
+    public function testUnauthenticatedUserCannotUnTagAPost()
+    {
+        // Create the models that we will be using
+        $post = factory(Post::class)->create();
+
+        // Apply the tag to the post
+        $response = $this->deleteJson('api/v3/posts/' . $post->id . '/tags', [
+            'tag_name' => 'Good Photo',
+        ]);
+
+        $response->assertStatus(401);
     }
 
     /**
      * Test deleting one tag on a post only deletes that tag
      *
-     * POST /tags
+     * POST /posts/:post_id/tag
      * @return void
      */
     public function testAddMultipleTagsAndDeleteOne()
     {
-        $this->actingAsAdmin();
+        // Create a post with tags.
+        // @TODO: Gateway keeps the "Token" from this PHPUnit call for later,
+        // and so we always think requests are anonymous. That's no good!
+        // We can swap this back once that's fixed in Gateway.
+        // $post = factory(Post::class)->create();
+        // $post->tag('Good Photo');
+        // $post->tag('Tag To Delete');
 
-        // Create a post with a tag.
         $post = factory(Post::class)->create();
-        $post->tag('Good Photo');
-        $post->tag('Tag To Delete');
+
+        $this->withAccessToken($this->randomUserId(), 'admin')->postJson('api/v3/posts/' . $post->id . '/tags', [
+             'tag_name' => 'Good Photo',
+         ]);
+
+        $this->withAdminAccessToken()->postJson('api/v3/posts/' . $post->id . '/tags', [
+             'tag_name' => 'Tag To Delete',
+         ]);
 
         // Make sure both tags actually exist
         $this->assertContains('Good Photo', $post->tagNames());
         $this->assertContains('Tag To Delete', $post->tagNames());
 
         // Send request to remove "Tag To Delete" tag
-        $response = $this->postJson('tags', [
-            'post_id' => $post->id,
+        $response = $this->withAccessToken($this->randomUserId(), 'admin')->deleteJson('api/v3/posts/' . $post->id . '/tags', [
             'tag_name' => 'Tag To Delete',
         ]);
 
@@ -97,28 +187,7 @@ class TagsTest extends TestCase
         $this->assertContains('Good Photo', $post->tagNames());
         $this->assertNotContains('Tag To Delete', $post->fresh()->tagNames());
 
-        // Make sure we created an event for the tag.
-        $this->assertDatabaseHas('events', [
-            'eventable_type' => 'Rogue\Models\Post',
-        ]);
-    }
-
-    /**
-     * Test that non-admin cannot tag or un-tag posts.
-     *
-     * @return void
-     */
-    public function testUnauthenticatedUserCantTag()
-    {
-        $user = factory(User::class)->create();
-        $post = factory(Post::class)->create();
-
-        $response = $this->actingAs($user)->postJson('tags', [
-            'post_id' => $post->id,
-            'tag_name' => 'Good Photo',
-        ]);
-
-        $response->assertStatus(403);
+        // @TODO: When we refactor events, make sure we created an event for the tag that was deleted.
     }
 
     /**
@@ -134,8 +203,7 @@ class TagsTest extends TestCase
         // Later, apply the tag to the post
         $this->mockTime('10/21/2017 13:05:00');
 
-        $this->actingAsAdmin()->postJson('tags', [
-            'post_id' => $post->id,
+        $this->withAccessToken($this->randomUserId(), 'admin')->postJson('api/v3/posts/' . $post->id . '/tags', [
             'tag_name' => 'Good Photo',
         ]);
 
@@ -143,7 +211,7 @@ class TagsTest extends TestCase
     }
 
     /**
-     * Test post updated_at is updated when a new tag is applied to it
+     * Test withoutTag scope
      *
      * @return void
      */
@@ -153,8 +221,7 @@ class TagsTest extends TestCase
         $posts = factory(Post::class, 20)->create();
 
         // Later, apply the tag to the post
-        $this->actingAsAdmin()->postJson('tags', [
-            'post_id' => $posts->first()->id,
+        $this->withAccessToken($this->randomUserId(), 'admin')->postJson('api/v3/posts/' . $posts->first()->id . '/tags', [
             'tag_name' => 'get-outta-here',
         ]);
